Dev: Included render to map, error Img dont apear!
